requirements.txt
alpaca_trade_api==3.2.0
Flask==3.0.2
Flask_SocketIO==5.3.6
python-dotenv==1.0.1
websocket_client==0.58.0
websocket_client==1.5.1
gunicorn==20.1.0  ## Required for deployment